+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased] >
 <!-- Separate headings for Added/Changed/Removed/Fixed/Deprecated/Security -->
+## [Unreleased]
+### Added
+- Added additional option `Y_col_block_size` to `MTLassoCV_MatchSpace_factory` to estimate `V` on block-averages of `Y` (e.g. taking a 150 cols down to 5 by doing averages over 30 cols at a time).
+- Added `se_factor` to `MTLassoCV_MatchSpace_factory` to use a different penalty than the MSE min.
 
 ## 0.2.0 - 2020-05-06
 ### Added